feat(core): wip update expression input parser
