+++ .vscode/settings.json
@@ -12,5 +12,6 @@
     "typedorm",
     "typeorm"
   ],
-  "editor.formatOnSave": true
+  "editor.formatOnSave": true,
+  "typescript.tsdk": "node_modules/typescript/lib"
 }
