+++ packages/common/src/operator-types.ts
@@ -18,6 +18,8 @@ export enum ATTRIBUTE_TYPE {
  */
 export type ArithmeticOperator = 'INCREMENT_BY' | 'DECREMENT_BY';
 
+export type SetUpdateOperator = 'IF_NOT_EXISTS' | 'APPEND_TO_LIST';
+
 /**
  * This is the single place where all the the document client supported
  * operators are listed
