+++ packages/core/src/classes/expression/projection-keys-options-type.ts
@@ -1,0 +1,1 @@
+export type ProjectionKeys<Entity> = (keyof Entity)[] | string[];
