+++ packages/core/src/classes/transformer/__test__/dc-request-transformer.spec.ts
@@ -666,7 +666,7 @@ test('transforms update item with primary key changes', () => {
       id: '1',
     },
     {
-      id: '1a',
+      id: '2',
     }
   );
